Marketplace: move product-install orchestration into a hook
Lift the state, effects, and error rendering out of the component and into a
useProductInstall hook, leaving the component as the page shell. Pure code
motion: the extracted body is unchanged, so the render, effects, and their
order are identical. This makes the orchestration unit-testable and sets up
splitting it per flow.
